+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Correct `normalize_rect` (oops!)
 - Add and document `finalize` method on ContentObjects
 - Make `PageList` work more or less like a `Sequence`
+- Support iteration over `playa.structure.ContentItem`
 - TODO: Maximize test coverage
 - TODO: Optimize marked content section access
 - TODO: Add method to complete parent tree for page